refactor: undo contract abi move
